+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/CenterRestController.java
@@ -27,4 +27,10 @@ public class CenterRestController {
         return centerController.createCenter(centerDTO);
     }
 
+    @PutMapping("/{centerId}")
+    public Center updateCenter(@PathVariable Long centerId, @RequestBody CenterDTO centerDTO) {
+        return centerController.updateCenter(centerId, centerDTO);
+    }
+
+
 }