+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/SchoolRestController.java
@@ -1,0 +1,15 @@
+package com.nimio2024.nimio2024_tfg_polmorenocastelar.application;
+
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers.SchoolController;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/schools")
+public class SchoolRestController {
+
+    SchoolController schoolController;
+
+    public SchoolRestController(SchoolController schoolController) {this.schoolController = schoolController;}
+
+}