+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/dto/CenterDTO.java
@@ -1,10 +1,15 @@
 package com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Center;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.School;
+
+import java.util.List;
 
 public class CenterDTO {
 
     private String centerName;
+
+    private List<School> schoolList;
 
     public CenterDTO() {
     }
@@ -17,6 +22,10 @@ public class CenterDTO {
         this.centerName = center.getCenterName();
     }
 
+    public CenterDTO(String centerName, List<School> schoolList){
+        this.centerName = centerName;
+        this.schoolList = schoolList;
+    }
     public String getCenterName() {
         return centerName;
     }
@@ -24,4 +33,6 @@ public class CenterDTO {
     public void setCenterName(String centerName) {
         this.centerName = centerName;
     }
+
+    public List<School> getSchoolList() {return schoolList;}
 }