+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/CenterController.java
@@ -25,4 +25,10 @@ public class CenterController {
         Center center = new Center(centerDTO);
         return centerService.saveCenter(center);
     }
+
+    public Center updateCenter(Long centerId, CenterDTO centerDTO) {
+        Center center = centerService.getCenterById(centerId);
+        center.setCenterName(centerDTO.getCenterName());
+        return centerService.saveCenter(center);
+    }
 }