+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/SchoolController.java
@@ -1,0 +1,14 @@
+package com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers;
+
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.service.SchoolService;
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class SchoolController {
+
+    SchoolService schoolService;
+
+    public SchoolController(SchoolService schoolService) {
+        this.schoolService = schoolService;
+    }
+}