+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/persistence/SchoolRepository.java
@@ -1,0 +1,9 @@
+package com.nimio2024.nimio2024_tfg_polmorenocastelar.persistence;
+
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Center;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SchoolRepository extends JpaRepository<Center, Long> {
+}