+++ NIMIO2024_TFG_PolMorenoCastelar/changelog.md
@@ -14,3 +14,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - [NPP-1] 》Create Spring Project 
 - [NPP-3] 》Architecture configuration & First Classes Testing
+- [NPP-9] 》CE1F - Create a Center and edit its Name